backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-dotenv
supabase
postgrest
bcrypt
numpy
google-generativeai
scikit-learn
joblib
